[CMake] Don't override -march or -mtune
Overriding compilation architecture settings is problematic because the
library does not know how it is intended to be used. This leads to the
following problems:

- The compilation and runtime instruction sets must match exactly.
Compiling the library on a host with e.g. AVX512 will cause the library
to crash on hosts without AVX512 and so on.

- On Apple platforms there exist hardware and software combinations
(e.g. M1 Mac and Xcode 12.x) where -march=native cannot be used due to
LLVM definition being not present yet.